pyproject.toml
[project]
name = "serial-monitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Max Herbert", email = "max@maxherbert.com" }]
requires-python = ">=3.11"
dependencies = [
    "textual>=5.2.0",
]

[project.scripts]
serial-monitor = "serial_monitor.main:main"

[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pyright>=1.1.403",
    "pytest>=8.4.1",
    "ruff>=0.12.7",
    "pyserial>=3.5",
    "textual-dev>=1.7.0",
    "black>=25.1.0",
]
